backend/agent/pyproject.toml
[project]
name = "text-diagrams-agent"
version = "0.1.0"
description = "Hybrid DSPy + FastAgent diagram generation agent"
requires-python = ">=3.13"
authors = [
    { name = "Text Diagrams Team" }
]

dependencies = [
    "dspy>=2.5.0",
    "fast-agent-mcp>=0.4.39",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=0.25",
    "pytest-mock>=3.14",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)
